onnx  icon

+++ YoloOnnxWinform/YoloOnnxWinform/FormYoloDetect.cs
@@ -1,10 +1,9 @@
 ﻿using CommImageControl;
-//using Compunet.YoloSharp;
-//using Compunet.YoloSharp.Plotting;
+
 using SkiaSharp;
 
-//using SixLabors.ImageSharp;
-using SkiaSharp;
+using SixLabors.ImageSharp;
+//using SkiaSharp;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -14,11 +13,15 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
-using YoloDotNet;
-using YoloDotNet.Core;
-using YoloDotNet.Enums;
-using YoloDotNet.Models;
-using YoloDotNet.Extensions;
+//using YoloDotNet;
+//using YoloDotNet.Core;
+//using YoloDotNet.Enums;
+//using YoloDotNet.Models;
+//using YoloDotNet.Extensions;
+
+using Compunet.YoloSharp;
+using Compunet.YoloSharp.Plotting;
+
 
 
 namespace YoloOnnxWinform
@@ -30,9 +33,9 @@ namespace YoloOnnxWinform
         public DataGridView DataGridList => this.dataGridView1;
         private System.Diagnostics.Stopwatch _stopwatch = new System.Diagnostics.Stopwatch();
 
-        //private YoloPredictor _yoloPredictor;
-        private DetectionDrawingOptions _drawingOptions;
-        private Yolo yoloPredictor;
+        private YoloPredictor _yoloPredictor;
+        //private DetectionDrawingOptions _drawingOptions;
+        //private Yolo yoloPredictor;
         public FormYoloDetect()
         {
             InitializeComponent();
@@ -41,33 +44,35 @@ namespace YoloOnnxWinform
 
         private void FormYoloDetect_Load(object sender, EventArgs e)
         {
-            // _yoloPredictor = new YoloPredictor("yolo11n.onnx");
+            _yoloPredictor = new YoloPredictor("best.onnx");
+            _yoloPredictor.Configuration.SuppressParallelInference = true;
+            _yoloPredictor.Configuration.KeepAspectRatio = true;
+            _yoloPredictor.Configuration.Confidence = 0.3F;
+            //_drawingOptions = new DetectionDrawingOptions
+            //{
+            //    DrawBoundingBoxes = true,
+            //    DrawConfidenceScore = true,
+            //    DrawLabels = true,
+            //    EnableFontShadow = true,
 
-            _drawingOptions = new DetectionDrawingOptions
-            {
-                DrawBoundingBoxes = true,
-                DrawConfidenceScore = true,
-                DrawLabels = true,
-                EnableFontShadow = true,
+            //    Font = SKTypeface.Default,
 
-                Font = SKTypeface.Default,
+            //    FontSize = 18,
+            //    FontColor = SKColors.White,
+            //    DrawLabelBackground = true,
+            //    EnableDynamicScaling = true,
+            //    BorderThickness = 2,
 
-                FontSize = 18,
-                FontColor = SKColors.White,
-                DrawLabelBackground = true,
-                EnableDynamicScaling = true,
-                BorderThickness = 2,
+            //    BoundingBoxOpacity = 128,
 
-                BoundingBoxOpacity = 128,
-
-            };
-            yoloPredictor = new Yolo(new YoloOptions
-            {
-                OnnxModel = "yolo11n.onnx",
-                ExecutionProvider = new CudaExecutionProvider(GpuId: 0, PrimeGpu: true),
-                ImageResize = ImageResize.Proportional,
-                SamplingOptions = new(SKFilterMode.Nearest, SKMipmapMode.None) // YoloDotNet default
-            });
+            //};
+            //yoloPredictor = new Yolo(new YoloOptions
+            //{
+            //    OnnxModel = "best.onnx",
+            //    ExecutionProvider = new CudaExecutionProvider(GpuId: 0, PrimeGpu: true),
+            //    ImageResize = ImageResize.Proportional,
+            //    SamplingOptions = new(SKFilterMode.Nearest, SKMipmapMode.None) // YoloDotNet default
+            //});
         }
 
         private void btnSelectDir_Click(object sender, EventArgs e)
@@ -121,7 +126,7 @@ namespace YoloOnnxWinform
             this.progressBar1.Value = 0;
             Task.Run(() =>
             {
-                _viewPresenter.Process(yoloPredictor);
+                _viewPresenter.Process(_yoloPredictor);
             });
         }
 
@@ -218,48 +223,48 @@ namespace YoloOnnxWinform
                 var item = _viewPresenter.GetSelectRowData(row);
                 if (item != null)
                 {
-                    //var result = _yoloPredictor.Detect(item.FilePath);
-                    //using var image = SixLabors.ImageSharp.Image.Load(item.FilePath);
+                    var result = _yoloPredictor.Detect(item.FilePath);
+                    using var image = SixLabors.ImageSharp.Image.Load(item.FilePath);
 
-                    //using var plot = result.PlotImage(image);
-                    //if (plot != null)
+                    using var plot = result.PlotImage(image);
+                    if (plot != null)
+                    {
+                        string folder = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "temp");
+                        if (!Directory.Exists(folder))
+                        {
+                            Directory.CreateDirectory(folder);
+                        }
+                        string path = Path.Combine(folder, item.FileName);
+                        if (File.Exists(path))
+                        {
+                            File.Delete(path);
+                        }
+                        plot.Save(path);
+                        FormUtils.Show(item.FileName, path);
+                    }
+
+                    //using var image = SKBitmap.Decode(item.FilePath);
+
+                    //// Run object detection inference
+                    //var results = yoloPredictor.RunObjectDetection(image, confidence: 0.15, iou: 0.7);
+
+                    //// Draw results
+                    //image.Draw(results, _drawingOptions);
+
+                    //string folder = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "temp");
+                    //if (!Directory.Exists(folder))
                     //{
-                    //    string folder = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "temp");
-                    //    if (!Directory.Exists(folder))
-                    //    {
-                    //        Directory.CreateDirectory(folder);
-                    //    }
-                    //    string path = Path.Combine(folder, item.FileName);
-                    //    if (File.Exists(path))
-                    //    {
-                    //        File.Delete(path);
-                    //    }
-                    //    plot.Save(path);
-                    //    FormUtils.Show(item.FileName, path);
+                    //    Directory.CreateDirectory(folder);
+                    //}
+                    //string path = Path.Combine(folder, item.FileName);
+                    //if (File.Exists(path))
+                    //{
+                    //    File.Delete(path);
                     //}
 
-                    using var image = SKBitmap.Decode(item.FilePath);
-
-                    // Run object detection inference
-                    var results = yoloPredictor.RunObjectDetection(image, confidence: 0.15, iou: 0.7);
-
-                    // Draw results
-                    image.Draw(results, _drawingOptions);
-
-                    string folder = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "temp");
-                    if (!Directory.Exists(folder))
-                    {
-                        Directory.CreateDirectory(folder);
-                    }
-                    string path = Path.Combine(folder, item.FileName);
-                    if (File.Exists(path))
-                    {
-                        File.Delete(path);
-                    }
-
-                    // Save image
-                    image.Save(path, SKEncodedImageFormat.Jpeg, 80);
-                    FormUtils.Show(item.FileName, path);
+                    //// Save image
+                    //image.Save(path, SKEncodedImageFormat.Jpeg, 80);
+                    //FormUtils.Show(item.FileName, path);
                 }
             }
             catch (Exception ex)
@@ -287,10 +292,10 @@ namespace YoloOnnxWinform
                 return;
             }
 
-            //if (_yoloPredictor != null)
-            //{
-            //    _yoloPredictor.Dispose();
-            //}
+            if (_yoloPredictor != null)
+            {
+                _yoloPredictor.Dispose();
+            }
         }
     }
 }
